+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#21](https://github.com/matheus-meneses/grafana-dynamic-search/issues/21): Match highlighting in dropdown results
 - [#75](https://github.com/matheus-meneses/grafana-dynamic-search/issues/75): Support dashboard variables for datasource selection
 
+### Fixed
+
+- [#77](https://github.com/matheus-meneses/grafana-dynamic-search/issues/77): Filter datasource variables to only show Prometheus-type variables
+
 ### Dependencies
 
 - Bump @swc/core from 1.15.8 to 1.15.10